# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "APL54-ESL Mock API",
    "version": "1.0.0",
    "description": "Mock ESL API for third-party client integration (POS / Inventory / Pricing systems)."
  },
  "servers": [
    {
      "url": "/",
      "description": "Dynamic (current host)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "VALIDATION_ERROR" },
              "message": { "type": "string", "example": "Human readable summary" },
              "details": { "type": "object", "example": { "field": "reason" } }
            }
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "refresh_token": { "type": "string" },
          "expires_in": { "type": "integer", "example": 300 },
          "token_type": { "type": "string", "example": "Bearer" }
        }
      },
      "Client": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "Your Company" }
        }
      },
      "Site": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "Store-1" },
          "location": { "type": "string", "example": "Demo Location" }
        }
      },
...
      "Template": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "price-barcode-184x88" },
          "display_type": { "type": "string", "example": "184x88" },
          "version": { "type": "integer", "example": 1 },
          "is_active": { "type": "boolean", "example": true }
        }
      },
...
      "JobRef": {
        "type": "object",
        "properties": {
          "job_id": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "example": "pending" },
          "total_count": { "type": "integer", "example": 1 }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "page_size": { "type": "integer", "example": 20 },
          "total": { "type": "integer", "example": 42 }
        }
      }
    }
  },
  "paths": {
    "/oauth/token": {
      "post": {
        "tags": ["Auth"],
        "summary": "Get or refresh access token",
        "description": "Use `client_credentials` to get a new token pair, or `refresh_token` to renew an access token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": ["grant_type"],
                "properties": {
                  "grant_type": {
                    "type": "string",
                    "enum": ["client_credentials", "refresh_token"],
                    "description": "OAuth2 grant type"
                  },
                  "client_id": {
                    "type": "string",
                    "description": "Required when grant_type=client_credentials"
                  },
                  "client_secret": {
                    "type": "string",
                    "description": "Required when grant_type=client_credentials"
                  },
                  "refresh_token": {
                    "type": "string",
                    "description": "Required when grant_type=refresh_token"
                  }
                }
              },
              "examples": {
                "client_credentials": {
                  "summary": "New token via client credentials",
                  "value": "grant_type=client_credentials&client_id=YOUR_CLIENT_ID&client_secret=YOUR_SECRET"
                },
                "refresh_token": {
                  "summary": "Refresh existing token",
                  "value": "grant_type=refresh_token&refresh_token=YOUR_REFRESH_TOKEN"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token issued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "401": {
            "description": "Unauthorized / invalid refresh token",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": ["Discovery"],
        "summary": "Get client info and sites",
        "description": "Returns your client record and all sites you have access to. Use this to discover your `site_id` values.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "client": { "$ref": "#/components/schemas/Client" },
                        "sites": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Site" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/v1/sites": {
      "get": {
        "tags": ["Discovery"],
        "summary": "List sites",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Site" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/v1/templates": {
      "get": {
        "tags": ["Templates"],
        "summary": "List templates for a site",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Site ID to list templates for"
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Template" }
                    },
                    "meta": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": {
            "description": "site_id not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/v1/templates/{template_id}/preview": {
      "post": {
        "tags": ["Templates"],
        "summary": "Preview a template with sample payload",
        "description": "Renders the template with given payload data and returns a base64-encoded PNG image. Use this to validate the display layout before sending an ESL update.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["site_id", "payload"],
                "properties": {
                  "site_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Site ID"
                  },
                  "payload": {
                    "type": "object",
                    "description": "Template variable data for preview",
                    "properties": {
                      "currency": { "type": "string", "example": "$" },
                      "price": { "type": "number", "example": 99.99 },
                      "product_id": { "type": "string", "example": "1234567890123" }
                    }
                  }
                }
              },
              "example": {
                "site_id": "uuid",
                "payload": {
                  "currency": "$",
                  "price": 99.99,
                  "product_id": "1234567890123"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preview image returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "template_id": { "type": "string", "format": "uuid" },
                        "content_type": { "type": "string", "example": "image/png" },
                        "image_base64": { "type": "string", "description": "Base64-encoded PNG image" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": {
            "description": "Template or site not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/v1/templates/{template_id}/activate": {
      "post": {
        "tags": ["Templates"],
        "summary": "Activate a template for a site",
        "description": "Activates the template for the given site. Only one template per display_type can be active at a time within a site.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["site_id"],
                "properties": {
                  "site_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Site ID to activate template for"
                  }
                }
              },
              "example": {
                "site_id": "uuid"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template activated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "template_id": { "type": "string", "format": "uuid" },
                        "display_type": { "type": "string", "example": "184x88" },
                        "is_active": { "type": "boolean", "example": true },
                        "activated_at": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": {
            "description": "Template or site not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
    "/api/v1/esl/{id}/reset": {
      "post": {
        "tags": ["ESL Tags"],
        "summary": "Reset ESL tag (device reset)",
        "description": "Sends a reset command to the ESL tag via its assigned AP. This is asynchronous — track progress with GET /jobs/{job_id}. Does NOT delete the tag from the database. Returns 409 if the tag has no assigned AP.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["site_id"],
                "properties": {
                  "site_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Site ID the tag belongs to"
                  },
                  "reason": {
                    "type": "string",
                    "nullable": true,
                    "description": "Optional audit reason for the reset",
                    "example": "device swap"
                  }
                }
              },
              "example": {
                "site_id": "uuid",
                "reason": "device swap"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Reset job accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/JobRef" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": {
            "description": "ESL tag or site not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "409": {
            "description": "AP not assigned to tag",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
  }
}